블랙시스트 조회 -> BlackList 클래스로 분리
